requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.1.0
billiard==4.2.1
celery==5.3.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-pkg==0.9
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.5
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.26.4
elastic-transport==8.17.0
elasticsearch==8.10.1
elasticsearch-dsl==8.9.0
idna==3.10
inflection==0.5.1
infobip-api-python-client==5.0.0
infobip-api-python-sdk==5.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
packaging==24.2
PaycomUz==3.1
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2==2.9.7
pydantic==1.10.21
pydantic_core==2.27.2
PyJWT==2.9.0
python-click==0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-environ==0.4.54
pytz==2025.1
PyYAML==6.0.2
redis==4.6.0
referencing==0.36.2
regex==2022.3.15
requests==2.31.0
rpds-py==0.23.1
setuptools==75.8.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.9.0
